server/world/chunk: fix unbounded loop in SubChunk.Layer and layer renumbering in compact
Neither of these is reachable from a packet, a world save or in-game play.
Both are reachable from the package's own API, which is exported.

SubChunk.Layer compared uint8(len(sub.storages)) against the layer index. At
256 storages that narrowing wraps to 0, which is <= every uint8, so the loop
never terminated and appended a 4096-entry storage until the process ran out
of memory. Layer(255) on a fresh sub chunk is enough; layer 254 terminates.
Every SetBlock call in dragonfly passes a literal layer of 0 or 1, and the
decode path never calls Layer at all, so only a caller outside the package can
reach it.

A sub chunk's storage count is encoded as a single byte in both the network
and disk formats, which also caps a decoded sub chunk at 255 storages. A 256th
storage cannot be represented at all: encoding one would write a count of 0
followed by 256 storages of payload. Layer now rejects layers at or above the
new MaxLayers rather than building a sub chunk that cannot be written out. The
three other places that narrowed a length to compare it against a layer index
are compared as ints for the same reason.

SubChunk.compact dropped every all-air storage and closed the gap, which moved
each surviving layer down. Layer numbers are semantic: layer 0 is the block
and layer 1 is Bedrock's waterlogging layer, so a sub chunk with air in layer
0 and water in layer 1 came out with the water in layer 0. Waterlogging always
writes a block to layer 0, and removing that block promotes the liquid back to
layer 0, so a chunk built through Tx never holds that state; it is reachable
through SetOpts.DisableLiquidDisplacement and through a Structure that returns
a liquid with no block. Only trailing all-air storages carry no information,
so only those are dropped now.

+++ server/world/chunk/chunk.go
@@ -102,7 +102,7 @@ func (chunk *Chunk) Sub() []*SubChunk {
 // sub chunk exists at the given y, the block is assumed to be air.
 func (chunk *Chunk) Block(x uint8, y int16, z uint8, layer uint8) uint32 {
 	sub := chunk.SubChunk(y)
-	if sub.Empty() || uint8(len(sub.storages)) <= layer {
+	if sub.Empty() || len(sub.storages) <= int(layer) {
 		return chunk.air
 	}
 	return sub.storages[layer].At(x, uint8(y), z)
@@ -112,7 +112,7 @@ func (chunk *Chunk) Block(x uint8, y int16, z uint8, layer uint8) uint32 {
 // SubChunk exists at the given y, a new SubChunk is created and the block is set.
 func (chunk *Chunk) SetBlock(x uint8, y int16, z uint8, layer uint8, block uint32) {
 	sub := chunk.sub[chunk.SubIndex(y)]
-	if uint8(len(sub.storages)) <= layer && block == chunk.air {
+	if len(sub.storages) <= int(layer) && block == chunk.air {
 		// Air was set at n layer, but there were less than n layers, so there already was air there.
 		// Don't do anything with this, just return.
 		return
